EQ1-207 refractor creer offre de stage (now in 1 file)
